Move tests requiring Docker to a separate source set
Exclude these tests during execution by passing `-x dockerTest` to the
./gradlew command.
