+++ sqldelight-gradle-plugin/src/dockerTest/kotlin/app/cash/sqldelight/dialect/DialectIntegrationTests.kt
@@ -1,0 +1,46 @@
+package app.cash.sqldelight.dialect
+
+import com.google.common.truth.Truth
+import org.gradle.testkit.runner.GradleRunner
+import org.junit.Test
+import java.io.File
+
+class DialectIntegrationTests {
+
+  @Test fun integrationTestsMySql() {
+    val runner = GradleRunner.create()
+      .withCommonConfiguration(File("src/test/integration-mysql"))
+      .withArguments("clean", "check", "--stacktrace")
+
+    val result = runner.build()
+    Truth.assertThat(result.output).contains("BUILD SUCCESSFUL")
+  }
+
+  @Test fun integrationTestsMySqlSchemaDefinitions() {
+    val runner = GradleRunner.create()
+      .withCommonConfiguration(File("src/test/integration-mysql-schema"))
+      .withArguments("clean", "check", "--stacktrace")
+
+    val result = runner.build()
+    Truth.assertThat(result.output).contains("BUILD SUCCESSFUL")
+  }
+
+  @Test fun integrationTestsPostgreSql() {
+    val runner = GradleRunner.create()
+      .withCommonConfiguration(File("src/test/integration-postgresql"))
+      .withArguments("clean", "check", "--stacktrace")
+
+    val result = runner.build()
+    Truth.assertThat(result.output).contains("BUILD SUCCESSFUL")
+  }
+}
+
+private fun GradleRunner.withCommonConfiguration(projectRoot: File): GradleRunner {
+  File(projectRoot, "gradle.properties").writeText(
+    """
+      |org.gradle.jvmargs=-Xmx4g -XX:MaxMetaspaceSize=1g
+      |
+    """.trimMargin()
+  )
+  return withProjectDir(projectRoot)
+}
